refactor(assay): extract imports.rs module
