+++ src/commands/assay/internal/mod.rs
@@ -2,6 +2,8 @@
 //!
 //! Follows dependable-rust pattern: private modules with curated re-exports.
 
+mod imports;
 mod util;
 
+pub(super) use imports::{execute_importers, execute_imports};
 pub(super) use util::truncate;
